refactor(server): split ws/handler.rs into focused event modules (#352)
