+++ apps/server/src/ws/events/mod.rs
@@ -1,0 +1,6 @@
+//! Per-event-type handlers dispatched from the main WebSocket receive loop.
+
+pub(super) mod broadcast;
+pub(super) mod canvas;
+pub(super) mod dispatch;
+pub(super) mod voice;
